+++ commit/operator/config.py
@@ -3,4 +3,4 @@ model         = None
 nIC           = None
 nEC      	  = None
 nISO     	  = None
-build_locally = False
+build_dir	  = None