+++ extras/COMMIT_debugger/main.cxx
@@ -66,7 +66,7 @@ float                    GLYPHS_affine[3][3];
 /*----------------------------------------------------------------------------------------------------------------------------------*/
 int main(int argc, char** argv)
 {
-    TCLAP::CmdLine cmd("This tool allows one to display in a common 3D space all the objects (DWI data, streamlines etc...) used by COMMIT in order to spot possible incosistencies between the conventions of COMMIT and the software that generated the data, e.g. flip in some axes in the DWI data or in the peaks, spatial shift in the streamlines, whether the affine transformation was already applied to the data etc..", ' ', "1.1");
+    TCLAP::CmdLine cmd("This tool allows one to display in a common 3D space all the objects (DWI data, streamlines etc...) used by COMMIT in order to spot possible incosistencies between the conventions of COMMIT and the software that generated the data, e.g. flip in some axes in the DWI data or in the peaks, spatial shift in the streamlines, whether the affine transformation was already applied to the data etc..", ' ', "1.2");
 
     TCLAP::UnlabeledValueArg<string> argDWI(    "dwi","Filename of the DWI dataset [4D NIFTI]", true, "", "DWI", cmd );
     TCLAP::ValueArg<string>          argMAP(    "m", "map", "Background map [3D NIFTI]", false, "", "map", cmd );
@@ -190,7 +190,8 @@ int main(int argc, char** argv)
         std::regex reVersion("^VERSION: (.*)\\s*$");
         std::smatch reMatches;
 
-        if ( !std::regex_match(string(line), reMatches, reVersion) )
+        std::string str_line = string(line);
+        if ( !std::regex_match(str_line, reMatches, reVersion) )
         {
             // no header found, assume standards BVECTOR format
             SCHEME_version = 0;
@@ -226,12 +227,13 @@ int main(int argc, char** argv)
         float       x, y, z, b, G, D, d;
         while( fgets(line, 1000, pFile) )
         {
-            if( std::regex_match(string(line), reMatches, reEMPTY) )
+            std::string str_line = string(line);
+            if( std::regex_match(str_line, reMatches, reEMPTY) )
                 continue;   // skip empty lines
 
             if( SCHEME_version == 0 )
             {
-                if ( !std::regex_match(string(line), reMatches, reVERSION0) )
+                if ( !std::regex_match(str_line, reMatches, reVERSION0) )
                     throw "Wrong row format";
                 x = std::atof( reMatches[1].str().c_str() );
                 y = std::atof( reMatches[2].str().c_str() );
@@ -244,7 +246,7 @@ int main(int argc, char** argv)
             }
             else
             {
-                if ( !std::regex_match(string(line), reMatches, reVERSION1) )
+                if ( !std::regex_match(str_line, reMatches, reVERSION1) )
                     throw "Wrong row format";
                 x = std::atof( reMatches[1].str().c_str() );
                 y = std::atof( reMatches[2].str().c_str() );
@@ -540,8 +542,7 @@ int main(int argc, char** argv)
         if ( TRK_file.hdr.dim[0] != dim.x || TRK_file.hdr.dim[1] != dim.y || TRK_file.hdr.dim[2] != dim.z ||
              abs(TRK_file.hdr.voxel_size[0]-pixdim.x) > 1e-4 || abs(TRK_file.hdr.voxel_size[1]-pixdim.y) > 1e-4 || abs(TRK_file.hdr.voxel_size[2]-pixdim.z) > 1e-4 )
         {
-            COLOR_error( "The GEOMETRY does not match those of DWI images", "\t" );
-            return EXIT_FAILURE;
+            COLOR_warning( "The GEOMETRY does not match those of DWI images", "\t" );
         }
 
         TRK_skip = ceil( TRK_file.hdr.n_count / 25000.0 );